Update 03_01-biv_asreml.Rmd
comment and modification 
Modification fixed effect / as(trait):sex  instead of trait:sex 
add the direct correlation estimation!  Model is not converging correctly !
creation of the partitioning section  but does not work also ! 
Also may be a better idea to directly split the response variable by sex and created a 4x4 multivariate model.  I used this kind of model in my work !